mesa: implement glNamedBufferDataEXT entrypoint
Note that the Driver.BufferData hook takes a 'target' parameter
as an optimization hint of how to allocate the buffer. Drivers
seem to be able to handle passing an invalid target of '0',
as done here, by falling back to a default.

Since it's merely an optimization hint, as any buffer could be
bound to many different targets, regardless of what target it
was bound to when its storage was first specified, drivers
in general shouldn't keel over because of this regardless.