mesa: fix assert/error on cube faces in get_and_init_texture
This would've triggered on any use of the Texture*
variants in GL_EXT_direct_state_access with a cube map
face target.

It'd either assert(name=0), or would fail to match the target
of the texture object looked up (either now, if it was already
initialized, or later, when used with a non-DSA entrypoint).